Move short_interrupt_gate_descriptor to top level.
