+++ x86/src/main/public/x86/_64/interrupts.h
@@ -3,6 +3,7 @@
 #pragma once
 
 #include <x86/common.h>
+#include <x86/registers.h>
 #include <x86/_64/descriptor.h>
