+++ x86/test/interrupts/src/main/cpp/main.cpp
@@ -12,7 +12,6 @@
 namespace
 {
     using namespace x86;
-    using namespace x86::_32;;
 
     struct
     {
@@ -55,7 +54,6 @@ void psys::main ()
 {
     using namespace ps;
     using namespace x86;
-    using namespace x86::_32;
 
     // set the GDT register and set segment registers
 
@@ -65,60 +63,10 @@ void psys::main ()
 
     set_segment_registers(segment_selector(2, false, 0), segment_selector(1, false, 0));
 
-    // test: data structures
-
     auto interrupt_segment = segment_selector(2, false, 0);
 
     for (auto i = 0U, j = 256U; i != j; ++i) {
-        interrupt_descriptor_table[i] = { interrupt_segment, interrupt_handler, true, true, 0, true };
-    }
-
-    auto& descriptor = interrupt_descriptor_table[0];
-
-    _test_control = 10;
-    if (descriptor.type() != 0b1110) {
-        _test_debug = descriptor.type();
-        _test_control = 0;
-        return;
-    }
-
-    _test_control = 11;
-    if (descriptor.privilege() != 0) {
-        _test_debug = descriptor.privilege();
-        _test_control = 0;
-        return;
-    }
-
-    _test_control = 12;
-    if (descriptor.is_present() != true) {
-        _test_control = 0;
-        return;
-    }
-
-    _test_control = 13;
-    if (descriptor.segment() != interrupt_segment) {
-        _test_debug = size2{descriptor.segment()};
-        _test_control = 0;
-        return;
-    }
-
-    _test_control = 14;
-    if (descriptor.offset() != halt_cast<size4>(interrupt_handler)) {
-        _test_debug = descriptor.offset();
-        _test_control = 0;
-        return;
-    }
-
-    _test_control = 15;
-    if (descriptor.is_32bit() != true) {
-        _test_control = 0;
-        return;
-    }
-
-    _test_control = 16;
-    if (descriptor.must_cli() != true) {
-        _test_control = 0;
-        return;
+        interrupt_descriptor_table[i] = { interrupt_segment, interrupt_handler, true, false, 0, true };
     }
 
     // set the IDT register
@@ -134,7 +82,7 @@ void psys::main ()
     auto expected_size = interrupt_descriptor_table_size * sizeof(short_interrupt_gate_descriptor);
     auto expected_offset = halt_cast<size4>(interrupt_descriptor_table);
 
-    auto [actual_size, actual_offset] = get_interrupt_descriptor_table();
+    auto [actual_size, actual_offset] = idtr();
 
     if (actual_size != expected_size || actual_offset != expected_offset) {
         _test_debug = expected_size;
